Add files via upload
SUPERVISED LEARNING (USING THE FAMOUS TITANIC DATASET)

For those who do not know about the titanic dataset; "it's a movie about a boat, the boat sink, people die! and  some survived" 

this is good dataset to practice our supervised technique, cause we have the answers (label) that we can use to train our model.